♻️ Add improvements (#5)
* 💄 Add TopBar

* 💄 Add El Reto de Hoy Cover

* 👔 Add Search

* 🚚 Add App icon

* 👽️ Dinamic fetch to El Reto de hoy Playlist

* 👽️ERDH Local Fetch

* 👽️ Clear Fetch ERDH Lists

* ♻️ Complete Refactor, all list locals and common views

---------
