+++ ibvn/Helper/Constants.swift
@@ -10,5 +10,7 @@ import SwiftUI
 struct Constants {
     // MARK: Colors
     static let primary = Color("Colors/primarySet")
+    static let primaryDark = Color("Colors/primaryDark")
     static let secondary = Color("Colors/secondarySet")
+    static let secondaryDark = Color("Colors/secondaryDark")
 }
