+++ ibvn/Modules/LocaData/NDVLists.json
@@ -1,0 +1,113 @@
+{
+    "kind": "youtube#searchListResponse",
+    "etag": "aY3iaClXeF2kxWi9Ay5fL7qdaD0",
+    "regionCode": "SV",
+    "pageInfo": {
+        "totalResults": 4,
+        "resultsPerPage": 3
+    },
+    "items": [
+        {
+            "kind": "youtube#searchResult",
+            "etag": "bViIHgcBVRIupvs6eS4owtW0c9g",
+            "id": {
+                "kind": "youtube#playlist",
+                "playlistId": "PL0MA6QqbTGP7-gDYhViwME7eag_9vr6Np"
+            },
+            "snippet": {
+                "publishedAt": "2024-01-09T22:07:08Z",
+                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
+                "title": "La Llave Bajo el Tapete - Levítico – Noche de Viernes",
+                "description": "En esta serie Steve Kern te ayudará a descubrir los secretos en el libro de Levítico para disfrutar de la presencia de Dios.",
+                "thumbnails": {
+                    "default": {
+                        "url": "https://i.ytimg.com/vi/KRciXi1d7cs/default.jpg",
+                        "width": 120,
+                        "height": 90
+                    },
+                    "medium": {
+                        "url": "https://i.ytimg.com/vi/KRciXi1d7cs/mqdefault.jpg",
+                        "width": 320,
+                        "height": 180
+                    },
+                    "high": {
+                        "url": "https://i.ytimg.com/vi/KRciXi1d7cs/hqdefault.jpg",
+                        "width": 480,
+                        "height": 360
+                    }
+                },
+                "channelTitle": "Iglesia Bautista Vida Nueva",
+                "liveBroadcastContent": "none",
+                "publishTime": "2024-01-09T22:07:08Z"
+            }
+        },
+        {
+            "kind": "youtube#searchResult",
+            "etag": "DMgZS41ytgAluEv8qZRMZ3r0NYY",
+            "id": {
+                "kind": "youtube#playlist",
+                "playlistId": "PL0MA6QqbTGP5EAhuojH80mJ2__4zBDjxB"
+            },
+            "snippet": {
+                "publishedAt": "2023-03-09T19:14:15Z",
+                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
+                "title": "El Origen de Todo - Genesis 1-3 - Noche de Viernes con Steve Kern",
+                "description": "",
+                "thumbnails": {
+                    "default": {
+                        "url": "https://i.ytimg.com/vi/8Y-0-l1yvco/default.jpg",
+                        "width": 120,
+                        "height": 90
+                    },
+                    "medium": {
+                        "url": "https://i.ytimg.com/vi/8Y-0-l1yvco/mqdefault.jpg",
+                        "width": 320,
+                        "height": 180
+                    },
+                    "high": {
+                        "url": "https://i.ytimg.com/vi/8Y-0-l1yvco/hqdefault.jpg",
+                        "width": 480,
+                        "height": 360
+                    }
+                },
+                "channelTitle": "Iglesia Bautista Vida Nueva",
+                "liveBroadcastContent": "none",
+                "publishTime": "2023-03-09T19:14:15Z"
+            }
+        },
+        {
+            "kind": "youtube#searchResult",
+            "etag": "4pMsuBU4FZLuC1ca2w6CH2AuX7E",
+            "id": {
+                "kind": "youtube#playlist",
+                "playlistId": "PL0MA6QqbTGP5wMa6jvEz8S0wvSLi8ivXB"
+            },
+            "snippet": {
+                "publishedAt": "2022-05-24T00:19:05Z",
+                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
+                "title": "Noches de VidaNueva - Viernes en vivo desde Campus Miralvalle",
+                "description": "",
+                "thumbnails": {
+                    "default": {
+                        "url": "https://i.ytimg.com/vi/itdf_fRNIJA/default.jpg",
+                        "width": 120,
+                        "height": 90
+                    },
+                    "medium": {
+                        "url": "https://i.ytimg.com/vi/itdf_fRNIJA/mqdefault.jpg",
+                        "width": 320,
+                        "height": 180
+                    },
+                    "high": {
+                        "url": "https://i.ytimg.com/vi/itdf_fRNIJA/hqdefault.jpg",
+                        "width": 480,
+                        "height": 360
+                    }
+                },
+                "channelTitle": "Iglesia Bautista Vida Nueva",
+                "liveBroadcastContent": "none",
+                "publishTime": "2022-05-24T00:19:05Z"
+            }
+        }
+    ]
+}
