--- ibvn/NDV/LocalData/NocheDeViernes.json
@@ -1,181 +1,0 @@
-{
-    "kind": "youtube#searchListResponse",
-    "etag": "TLdlVAyS2hsoSzZ5LpoZbwCv0FI",
-    "regionCode": "SV",
-    "pageInfo": {
-        "totalResults": 66,
-        "resultsPerPage": 5
-    },
-    "items": [
-        {
-            "kind": "youtube#searchResult",
-            "etag": "hJ6PHQtfbgpdVFglLinRQkR70wI",
-            "id": {
-                "kind": "youtube#video",
-                "videoId": "XUzjUC6sloQ"
-            },
-            "snippet": {
-                "publishedAt": "2024-03-23T15:03:36Z",
-                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
-                "title": "No Cruces la Línea con Dios - Levítico 7:11-38 / Steve Kern - #ndv",
-                "description": "Qué pasa si a Dios se le acaba la paciencia? La historia que descubrirás hoy, te mostrará que la mejor muestra de amor que ...",
-                "thumbnails": {
-                    "default": {
-                        "url": "https://i.ytimg.com/vi/XUzjUC6sloQ/default.jpg",
-                        "width": 120,
-                        "height": 90
-                    },
-                    "medium": {
-                        "url": "https://i.ytimg.com/vi/XUzjUC6sloQ/mqdefault.jpg",
-                        "width": 320,
-                        "height": 180
-                    },
-                    "high": {
-                        "url": "https://i.ytimg.com/vi/XUzjUC6sloQ/hqdefault.jpg",
-                        "width": 480,
-                        "height": 360
-                    }
-                },
-                "channelTitle": "Iglesia Bautista Vida Nueva",
-                "liveBroadcastContent": "none",
-                "publishTime": "2024-03-23T15:03:36Z"
-            }
-        },
-        {
-            "kind": "youtube#searchResult",
-            "etag": "XB6LzoNcr7mh9e5rNCHzZyNAw_E",
-            "id": {
-                "kind": "youtube#video",
-                "videoId": "20rCsebIMRQ"
-            },
-            "snippet": {
-                "publishedAt": "2023-07-17T06:05:21Z",
-                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
-                "title": "El Secreto para Comenzar una Vida Nueva - Lucas",
-                "description": "Muchos necesitamos la oportunidad de un nuevo comienzo para así tener una vida nueva. En este mensaje descubrirás que es ...",
-                "thumbnails": {
-                    "default": {
-                        "url": "https://i.ytimg.com/vi/20rCsebIMRQ/default.jpg",
-                        "width": 120,
-                        "height": 90
-                    },
-                    "medium": {
-                        "url": "https://i.ytimg.com/vi/20rCsebIMRQ/mqdefault.jpg",
-                        "width": 320,
-                        "height": 180
-                    },
-                    "high": {
-                        "url": "https://i.ytimg.com/vi/20rCsebIMRQ/hqdefault.jpg",
-                        "width": 480,
-                        "height": 360
-                    }
-                },
-                "channelTitle": "Iglesia Bautista Vida Nueva",
-                "liveBroadcastContent": "none",
-                "publishTime": "2023-07-17T06:05:21Z"
-            }
-        },
-        {
-            "kind": "youtube#searchResult",
-            "etag": "CgC40-NIJBG15VsB1yzomC3YD9o",
-            "id": {
-                "kind": "youtube#video",
-                "videoId": "ZJWCUBfEmic"
-            },
-            "snippet": {
-                "publishedAt": "2020-03-21T02:27:11Z",
-                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
-                "title": "Preguntas y Respuestas - 20 de marzo",
-                "description": "Haz tus preguntas por whatsapp al: (503) 6200-8558.",
-                "thumbnails": {
-                    "default": {
-                        "url": "https://i.ytimg.com/vi/ZJWCUBfEmic/default.jpg",
-                        "width": 120,
-                        "height": 90
-                    },
-                    "medium": {
-                        "url": "https://i.ytimg.com/vi/ZJWCUBfEmic/mqdefault.jpg",
-                        "width": 320,
-                        "height": 180
-                    },
-                    "high": {
-                        "url": "https://i.ytimg.com/vi/ZJWCUBfEmic/hqdefault.jpg",
-                        "width": 480,
-                        "height": 360
-                    }
-                },
-                "channelTitle": "Iglesia Bautista Vida Nueva",
-                "liveBroadcastContent": "none",
-                "publishTime": "2020-03-21T02:27:11Z"
-            }
-        },
-        {
-            "kind": "youtube#searchResult",
-            "etag": "gINojnp3cFOGuE4Pa2QKEMujgNA",
-            "id": {
-                "kind": "youtube#video",
-                "videoId": "DiqQOt_oo28"
-            },
-            "snippet": {
-                "publishedAt": "2019-03-19T15:47:06Z",
-                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
-                "title": "2019-03-08 - Preguntas y Respuestas NDV - David Aguilar",
-                "description": "",
-                "thumbnails": {
-                    "default": {
-                        "url": "https://i.ytimg.com/vi/DiqQOt_oo28/default.jpg",
-                        "width": 120,
-                        "height": 90
-                    },
-                    "medium": {
-                        "url": "https://i.ytimg.com/vi/DiqQOt_oo28/mqdefault.jpg",
-                        "width": 320,
-                        "height": 180
-                    },
-                    "high": {
-                        "url": "https://i.ytimg.com/vi/DiqQOt_oo28/hqdefault.jpg",
-                        "width": 480,
-                        "height": 360
-                    }
-                },
-                "channelTitle": "Iglesia Bautista Vida Nueva",
-                "liveBroadcastContent": "none",
-                "publishTime": "2019-03-19T15:47:06Z"
-            }
-        },
-        {
-            "kind": "youtube#searchResult",
-            "etag": "1RMheV-HfK_CeeuMo5HApgrFh-M",
-            "id": {
-                "kind": "youtube#video",
-                "videoId": "I2zDm3wy21M"
-            },
-            "snippet": {
-                "publishedAt": "2018-04-04T21:00:08Z",
-                "channelId": "UCoNq7HF7vnqalfg-lTaxrDQ",
-                "title": "2018-03-23 - RESPUESTAS DE LA BIBLIA - NDV",
-                "description": "",
-                "thumbnails": {
-                    "default": {
-                        "url": "https://i.ytimg.com/vi/I2zDm3wy21M/default.jpg",
-                        "width": 120,
-                        "height": 90
-                    },
-                    "medium": {
-                        "url": "https://i.ytimg.com/vi/I2zDm3wy21M/mqdefault.jpg",
-                        "width": 320,
-                        "height": 180
-                    },
-                    "high": {
-                        "url": "https://i.ytimg.com/vi/I2zDm3wy21M/hqdefault.jpg",
-                        "width": 480,
-                        "height": 360
-                    }
-                },
-                "channelTitle": "Iglesia Bautista Vida Nueva",
-                "liveBroadcastContent": "none",
-                "publishTime": "2018-04-04T21:00:08Z"
-            }
-        }
-    ]
-}
